backend_folder/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-model-utils
djongo
psycopg2
openai
django-environ
pymongo